requirements.txt
anyio==3.6.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cloudinary==1.29.0
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==4.0.5
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-registration-redux==2.10
django-tinymce==3.4.0
djangorestframework==3.13.1
dnspython==2.2.1
email-validator==1.2.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.3
importlib-metadata==4.11.4
makefile==1.1.0
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pydantic==1.9.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2021.3
pyuploadcare==3.0.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.9
whitenoise==6.2.0
zipp==3.8.0
